+++ frontend/src/app/globals.css
@@ -24,5 +24,5 @@ body {
   color: var(--foreground);
   background: var(--background);
   font-family: Arial, Helvetica, sans-serif;
-  height: 1080px;
+  /* height: 1080px; */
 }
