+++ frontend/src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import style from "@/app/footer.module.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -21,44 +20,6 @@ export const metadata: Metadata = {
   },
 };
 
-const footerData = [
-  {
-    lines: ["GPS System", "은성트래시스(주)"],
-  },
-  {
-    lines: [
-      "대표 : 양국승",
-      "· 개인정보 보호책임자 : 양국승",
-      "· 사업자번호 : 538 - 88 - 01948",
-      "· 통신판매업 신고번호 : 2022 - 광주북구 - 0425",
-    ],
-  },
-  {
-    lines: [
-      "사무실 : 62286 광주광역시 광산구 장신로 342 - 19",
-      "Copyright (C) 2024 은성트래시스(주) all rights reserved.",
-    ],
-  },
-];
-
-function Footer() {
-  return (
-    <div className={style.container}>
-      <div className={style.wrapper}>
-        {footerData.map((section, index) => (
-          <div key={index} className={style.first}>
-            {section.lines.map((line, idx) => (
-              <span key={idx} className={style.span}>
-                {line}
-              </span>
-            ))}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
 export default function RootLayout({
   children,
 }: Readonly<{
@@ -70,7 +31,7 @@ export default function RootLayout({
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         {children}
-        <Footer />
+        {/* <Footer /> */}
       </body>
     </html>
   );
